Destination - Video gallery - part 2

+++ Travel_Ageancy/app/Http/Controllers/Admin/AdminDestinationController.php
@@ -111,6 +111,10 @@ class AdminDestinationController extends Controller
         if($total > 0) {
             return redirect()->back()->with('error', 'First Delete All Photos of This Destination');
         }
+        $total1 = DestinationVideo::where('destination_id', $id)->count();
+        if($total1 > 0) {
+            return redirect()->back()->with('error', 'First Delete All Videos of This Destination');
+        }
         $destination = Destination::where('id', $id)->first();
         unlink(public_path('uploads/'.$destination->featured_photo));
         $destination->delete();
